restructure project for testing
