+++ pyproject.toml
@@ -6,7 +6,7 @@ readme = "README.md"
 requires-python = ">=3.13"
 license = {text = "MIT"}
 authors = [
-    {name = "Materialize", email = "support@materialize.com"}
+    {name = "Materialize Inc", email = "support@materialize.com"}
 ]
 classifiers = [
     "Development Status :: 4 - Beta",
@@ -29,7 +29,7 @@ dependencies = [
 ]
 
 [project.scripts]
-materialize-mcp = "materialize_mcp_server:run"
+materialize-mcp = "materialize_mcp_server:main"
 
 [project.urls]
 Homepage = "https://github.com/MaterializeInc/materialize-mcp-server"
@@ -43,3 +43,11 @@ build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
 packages = ["materialize_mcp_server"]
+
+[dependency-groups]
+dev = [
+    "docker>=7.1.0",
+    "pytest-asyncio>=0.26.0",
+    "pytest>=8.3.5",
+    "testcontainers>=4.10.0",
+]
